Make core drop-down toolbar logic reusable
This extracts the core drop-down toolbar logic from the audio track toolbar into a form where it can be reused. (It's not an ideal form, but sharing some of this logic isn't easy.)

There was also some other minor drop-down toolbar tidying up.
