+++ foo_ui_columns/audio_track_toolbar.cpp
@@ -1,110 +1,15 @@
 #include "pch.h"
 
+#include "core_drop_down_list_toolbar.h"
 #include "drop_down_list_toolbar.h"
 
 namespace cui {
 
 namespace {
 
-constexpr GUID stream_selector_api_id{0x6A6FF0B0, 0x3FB8, 0x4413, {0xBB, 0x76, 0xFF, 0x48, 0xC8, 0x8F, 0xF0, 0x57}};
-
-fb2k::toolbarDropDown::ptr get_stream_selector_api()
-{
-    for (auto api : fb2k::toolbarDropDown::enumerate()) {
-        if (api->getGuid() == stream_selector_api_id)
-            return api;
-    }
-
-    return {};
-}
-
-struct AudioTrackToolbarArgs {
-    using ID = size_t;
-    using ItemList = std::vector<std::tuple<ID, std::string>>;
-
-    class AudioTrackCallback : public fb2k::toolbarDropDownNotify {
-    public:
-        void contentChanged() override { DropDownListToolbar<AudioTrackToolbarArgs>::s_refresh_all_items_safe(); }
-        void selectionChanged() override { DropDownListToolbar<AudioTrackToolbarArgs>::s_update_active_item_safe(); }
-    };
-
-    static auto get_items()
-    {
-        ItemList items;
-
-        const auto api = get_stream_selector_api();
-
-        if (api.is_empty())
-            return items;
-
-        for (const auto index : std::ranges::views::iota(size_t{0}, api->getNumValues())) {
-            pfc::string8 value;
-            api->getValue(index, value);
-
-            if (stricmp_utf8(value, "not playing") != 0)
-                items.emplace_back(index, value);
-        }
-
-        return items;
-    }
-
-    static ID get_active_item()
-    {
-        const auto api = get_stream_selector_api();
-
-        if (api.is_empty())
-            return ID{};
-
-        return api->getSelectedIndex();
-    }
-
-    static void set_active_item(ID id)
-    {
-        const auto api = get_stream_selector_api();
-
-        if (api.is_empty())
-            return;
-
-        api->setSelectedIndex(id);
-    }
-
-    static void on_click()
-    {
-        const auto api = get_stream_selector_api();
-
-        if (api.is_empty())
-            return;
-
-        api->onDropDown();
-    }
-
-    static void get_menu_items(uie::menu_hook_t& p_hook) {}
-
-    static void on_first_window_created()
-    {
-        const auto api = get_stream_selector_api();
-
-        if (api.is_empty())
-            return;
-
-        s_callback = std::make_unique<AudioTrackCallback>();
-        api->addNotify(s_callback.get());
-    }
-
-    static void on_last_window_destroyed()
-    {
-        const auto api = get_stream_selector_api();
-
-        if (api.is_empty())
-            return;
-
-        api->removeNotify(s_callback.get());
-        s_callback.reset();
-    }
-
-    static bool is_available() { return get_stream_selector_api().is_valid(); }
-    inline static std::unique_ptr<AudioTrackCallback> s_callback;
-    static constexpr bool refresh_on_click = false;
+struct AudioTrackCoreToolbarArgs {
+    static constexpr GUID core_toolbar_id{0x6A6FF0B0, 0x3FB8, 0x4413, {0xBB, 0x76, 0xFF, 0x48, 0xC8, 0x8F, 0xF0, 0x57}};
+    static constexpr const char* ignored_value{"not playing"};
     static constexpr auto no_items_text = "(not playing)"sv;
     static constexpr const wchar_t* class_name{L"columns_ui_audio_track_toolbar_xvkMz8coqQY"};
     static constexpr const char* name{"Audio track"};
@@ -113,7 +18,7 @@ struct AudioTrackToolbarArgs {
     static constexpr GUID font_client_id{0xe547f854, 0x1efe, 0x4fca, {0x8d, 0x42, 0xe4, 0x24, 0x99, 0x12, 0xbc, 0x9a}};
 };
 
-ui_extension::window_factory<DropDownListToolbar<AudioTrackToolbarArgs>> _;
+ui_extension::window_factory<DropDownListToolbar<CoreDropDownToolbarArgs<AudioTrackCoreToolbarArgs>>> _;
 
 } // namespace
